Raise NotFoundError if the list_blobs returns an iterator of 0 blobs, as per reviewer's request.

+++ storage/google_storage.py
@@ -74,7 +74,12 @@ class GoogleStorage(Storage):
         bucket = self._get_bucket()
 
         prefix = self._parsed_storage_uri.path[1:] + "/"
-        for blob in bucket.list_blobs(prefix=prefix):
+
+        blobs = list(bucket.list_blobs(prefix=prefix))
+        if len(blobs) == 0:
+            raise NotFoundError("No Files Found")
+
+        for blob in blobs:
             relative_path = blob.name.replace(prefix, "", 1)
             local_file_path = os.path.join(directory_path, relative_path)
             local_directory = os.path.dirname(local_file_path)
@@ -87,7 +92,7 @@ class GoogleStorage(Storage):
                 try:
                     retry.attempt(unversioned_blob.download_to_filename, local_file_path)
                 except NotFound as original_exc:
-                    raise NotFoundError("No Files Found") from original_exc
+                    raise NotFoundError("No File Found") from original_exc
 
     def load_from_directory(self, directory_path: str) -> None:
         bucket = self._get_bucket()
@@ -104,9 +109,13 @@ class GoogleStorage(Storage):
     def delete_directory(self) -> None:
         bucket = self._get_bucket()
 
-        for blob in bucket.list_blobs(prefix=self._parsed_storage_uri.path[1:] + "/"):
+        blobs = list(bucket.list_blobs(prefix=self._parsed_storage_uri.path[1:] + "/"))
+        if len(blobs) == 0:
+            raise NotFoundError("No Files Found")
+
+        for blob in blobs:
             unversioned_blob = bucket.blob(blob.name)
             try:
                 unversioned_blob.delete()
             except NotFound as original_exc:
-                raise NotFoundError("No Files Found") from original_exc
+                raise NotFoundError("No File Found") from original_exc
